dashboard_data/scripts/requirements.txt
pyodbc>=5.1.0
python-dateutil>=2.9.0
python-dotenv>=1.0.1
requests>=2.32.2
